requirements.txt
numpy
opencv_python_headless
Pillow
streamlit
